add function to return original cover image url
as of PB v3.9.5.1, PB has a filter hook called 'pb_cover_image' to allow
users to alter the cover image url for those users who do not store images
on the local server. This new function hooks into that filter and returns
$metadata['pb_cover_image'] (the original image url) instead of the url
stored in $cover variable in class-pb-catalog.php.