Add IMSCC Import logic
This commit moves the IMSCC logic out of pressbooks core, in
anticipation of a new action/filter hook system that will allow
users to add custom imports of various file types:

pb_import_file_types is a filter that allows users to append a
custom import file type to the approved list of file types.

pb_initialize_import is an filter that supplies an argument to
setCurrentImportOption in the custom import type Class.

pb_select_import_type is a filter that allows users to add a
custom import type to the drop down select in the import menu