fix 'max-width' bug on catalog page
The Catalog side bar was not displaying the proper color for the
entire width of the bar, because the bar was getting set to a
max-width of 290px.  This commit adds a catalog css file that
sets the max-width to none, also adding the ability to add styles
to the catalog in the future.